requirements.txt
anthropic>=0.40
python-dotenv>=1.0.0
reportlab>=4.0.0
pytest>=8.0.0
flask>=3.0.0
twilio>=9.0.0